perf: add css-reset stylesheet
Move /*.css to /css/main.css, and @import /css/normalize.css

There are many preset properties that affect the style of the page; this
can lead to inconsistent presentation across browsers. To avoid this, a
CSS-reset stylesheet is used to normalize the browser. This stylesheet
is robust and tries to preserve useful defaults.

The CSS-reset stylesheet is open-source and, as of this commit date,
maintained at https://github.com/neoclas/normalize.css .

Prior to this commit, the margin and padding were normalized through CSS
resets in the main stylesheet ( /*.css ).