feat: conform to the apca bronze-level standard
Change "... { *color: ... ; font*: ... ; }" in /css/main.css

APCA standard, as of this commit date, at https://git.apcacontrast.com
defines a method and metric for improving visual readability. In this,
several levels of conformance are defined - bronze, silver, and gold.
There will be an effort to, at minimum, conform to the bronze-level.
This required changing colors for >75Lc and font properties for better
rendering.